bump: hyperon-runtime to 2.0.2.
Added java 11 compatibility. Attached zip file at README.md

+++ src/test/java/pl/decerto/hyperon/guide/domain/ComplexAttributeDomainObjectTest.java
@@ -192,7 +192,7 @@ class ComplexAttributeDomainObjectTest {
 		boolean isAttributePremiumSet = false;
 		String premiumCode = "";
 		String premiumName = "";
-		int premiumValue = 0;
+		double premiumValue = 0d;
 		Type rawType = null;
 		String rawValue = "";
